refactored but error
