+++ client/store/index.js
@@ -2,7 +2,7 @@ import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 
-import { productReduce, authReducer, reviewReducer } from './reducers';
+import { productReducer, authReducer, reviewReducer } from './reducers';
 
 const reducer = combineReducers({
   products: productReducer,
